TS-2: Fix compiler error by casting pthread_t to a long instead of an int. error: cast from '_opaque_pthread_t*' to 'unsigned int' loses precision
git-svn-id: https://svn.apache.org/repos/asf/incubator/trafficserver/traffic/trunk@884397 13f79535-47bb-0310-9956-ffa450edef68

+++ libinktomi++/Diags.cc
@@ -212,7 +212,7 @@ Diags::print_va(const char *debug_tag, DiagsLevel diags_level,
 
   // add the thread id
   pthread_t id = pthread_self();
-  end_of_format += snprintf(end_of_format, sizeof(format_buf), "{%u} ", (unsigned) id);
+  end_of_format += snprintf(end_of_format, sizeof(format_buf), "{%lu} ", (unsigned long) id);
 
   ////////////////////////////////
   // start with the user prefix //
